pyproject.toml
[project]
name = "nodesk"
version = "0.0.0"
description = "Backend - NoDesk - Pro4tech (API Fatec)"
authors = [{ name = "Jhonny Freitas Dutra", email = "dutrajy@gmail.com" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "sqlalchemy[mypy] (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "psycopg[binary] (>=3.2.10,<4.0.0)",
    "pydantic[email] (>=2.11.7,<3.0.0)",
    "httpx[cli] (>=0.28.1,<0.29.0)",
    "asgi-lifespan (>=2.1.0,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "types-passlib (>=1.7.7.20250602,<2.0.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.11"
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
httpx = "^0.28.1"
pre-commit = "^4.3.0"
mypy = "^1.17.1"

[tool.ruff]
target-version = "py313"
lint.select = ["E", "F", "I", "UP", "B", "SIM"]
exclude = [".venv"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.13"
plugins = ['pydantic.mypy', 'sqlalchemy.ext.mypy.plugin']
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_reexport = true
check_untyped_defs = true

[tool.alembic]
script_location = "nodesk/core/database/migrations"
prepend_sys_path = ["."]

[[tool.alembic.post_write_hooks]]
name = "ruff_format"
type = "exec"
executable = "ruff"
options = "format REVISION_SCRIPT_FILENAME"

[[tool.alembic.post_write_hooks]]
name = "ruff_fix"
type = "exec"
executable = "ruff"
options = "check --fix REVISION_SCRIPT_FILENAME"
